+++ etc/blocks/enter-numbers/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @milaboratories/milaboratories.test-enter-numbers
 
+## 1.0.81
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.81
+
+## 1.0.80
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.80
+
+## 1.0.79
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.79
+
 ## 1.0.78
 
 ### Patch Changes