+++ etc/uikit-playground/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @milaboratories/uikit-playground
 
+## 0.0.50
+
+### Patch Changes
+
+- Updated dependencies [dfa2765]
+  - @milaboratories/uikit@2.2.3
+
+## 0.0.49
+
+### Patch Changes
+
+- Updated dependencies [d85c758]
+  - @milaboratories/uikit@2.2.2
+
 ## 0.0.48
 
 ### Patch Changes